Updates
Migrations: Altered products and games to add a date modified column

jwt-auth: fixed issue with code. next() was not placed correctly which was causing an issue with the promise chain

admin-auth: added middleware to check for admin privledges on authorized routes

users, tickets, games: added middleware to check for admin privledges and added modified date to any update routes.